guile30: Update to 3.0.3
Changes in 3.0.3 (since 3.0.2)

[Bugfixes and minor improvements]

* New baseline compiler

Guile's CPS-based compiler generates good code, but it takes time and
memory to do so.  For users that prioritize speed of compilation over
speed of generated code, Guile now has a new baseline compiler that goes
directly from the high-level Tree-IL to bytecode, skipping CPS and all
of its optimizations.  This compiler is used for `guild compile -O0',
and generally runs around ten times as fast as the CPS compiler.

* Faster Guile build from source

Guile now uses the baseline compiler for its bootstrap, when building
the first Scheme compiler.  Because the baseline compiler runs faster
and includes less code than the CPS compiler, Guile takes less time to
build.

** Old bitvector interfaces deprecated

See "Bit Vectors" in the manual, for details on all of these
replacements.