barony337: Import Barony-3.3.7 as wip/barony337
Motivation for separate package:

This package is based on wip/barony. The game engine (package) version
needs to match the game data version (but not precisely), so create a
separate package for this version.

Description:

Barony is a 3D, first-person roguelike. The goal of the game is to
descend to the bottom of a dark dungeon known as the Devil's Bastion
and destroy an undead lich named Baron Herx, who terrorized the
peaceful town of Hamlet in life and is now harboring a curse against
the land from beyond the grave. To aid you in your quest are friendly
humans who have been eeking out a rough life within the dungeon
for generations, as well as any friends you can bring with you in
real life: Barony is the first of its kind as a first-person
roguelike in that it fully supports cooperative multiplayer for up
to four players.