riak14: Adjust erlang dependency (not tested)
It is intuitively obvious to the casual observer that with a checkout
of wip and pkgsrc HEAD, all references in wip should resolve, although
this only matters for lintpkgsrc sorts of things.  Repoint the erlang
dependency from wip/erlang15 to wip/erlang-basho, which is clearly no
worse and might well be better.