wip/py-tinyarray: import py37-tinyarray-1.2.2
Tinyarrays are similar to NumPy arrays, but optimized for small sizes. Common
operations on very small arrays are to 3-7 times faster than with NumPy (with
NumPy 1.6 it used to be up to 35 times), and 3 times less memory is used to
store them. Tinyarrays are useful if you need many small arrays of numbers,
and cannot combine them into a few large ones. (The resulting code is still
much slower than C, but it may now be fast enough.)