wip/py-symEnergy: import py37-symenergy-1.0.7
SymEnergy provides a framework for the structured solution of cost-optimization
of energy systems using the standard Lagrange multiplier approach. The result
consists in the close-form analytical solutions to the optimization problem.
For example, the energy production from a power plant is expressed as a
function of the vector of symbolic parameters. These solutions are evaluated
for certain parameters in order to identify relevant constraint combinations