rust: Import version 1.47.0.
This is a copy of lang/rust, updated to 1.47.0, but will track the
latest stable release of rust, for when either the lack of NetBSD
bootstraps or some package dependencies stop us from upgrading.

The alternative was to import it as lang/rust-latest, which didn't
really sit well with me.  Maybe in time if this continues to be an issue
we can import this as lang/rust and create a lang/rust-legacy or similar
for systems that can't update.