kermit: Update to pre-alpha03
This is very very WIP-y, more so than usual.

Besides the update:

* delete ancient patch with no effect

We added this in 2012, and the issue has apparently been addressed in
some other way.  Upstream has opined that it's not necessary, and
building without it did not exhibit the warning for which it was
added.

* Document a patch better

Note that the TERMIO regularizaiton is not appropriate for upstream.

* Drop patches Overtaken By Upstream

Upstream has revamped the strategy for declartion of errno, and thus
these are no longer needed.

* Drop errno-declaring patch

Upstream addressed this more broadly and our patch is not longer
needed.

* Drop errno.h include, rather than adding one

Likely will be applied upstream.