openblas: version 0.3.10
This updated fixes the build with gcc 10 (segfault in cblat1 test).

I did not go all the way to the current 0.3.12, as that would
need some hacking of chosen compiler flags. 0.3.13 should be
the next one.

This commit also fixes the ARCH → ARCH_ sed to change all
occurences on a line.