modular-xorg-xspice: import modular-xorg-xspice-0.1.5 as modular-xorg-xspice
Description:

The QXL virtual GPU is found in the RedHat Enterprise Virtualisation system,
and also in the spice project. This package provides the Xspice server script
and Xorg drivers.

Notes:

Connecting with spicy (sysutils/spice-gtk) seems to be quite slow. Using vdagent
is supposed to improve the situation, but I did not manage to get it to work
correctly. The package does not proviude documentation, see Xspice --help and

https://gitlab.freedesktop.org/xorg/driver/xf86-video-qxl/-/blob/master/README.xspice