Add a patch for NetBSD for the vendor/net2 crate.
NetBSD doesn't actually need the same workaround as OpenBSD and Haiku;
on the TCP level, the TCP_KEEPALIVE socket option is recognized.

Without this we saw

 13038     19 routinator CALL  setsockopt(0x15,0xffff,8,0x7d60bd3d62b4,4)
 13038     19 routinator RET   setsockopt 0
 13038     19 routinator CALL  setsockopt(0x15,6,8,0x7d60bd3d62b4,4)
 13038     19 routinator RET   setsockopt -1 errno 42 Protocol option not available

and '8' in the latter corresponds to the

from <netinet/tcp.h> if I'm not sorely mistaken.  Instead we end up
using TCP_KEEPIDLE, which works fine by the looks of it.