bats-core: Import bats-core-1.2.1 as wip/bats-core
Bats is a TAP-compliant testing framework for Bash. It provides a
simple way to verify that the UNIX programs you write behave as
expected.

A Bats test file is a Bash script with special syntax for defining
test cases. Under the hood, each test case is just a function with
a description.