ruby-asciidoctor: update to 2.0.12
v2.0.12 @mojavelinux released this on Nov 10 2020

This is a minor patch release that fixes some regressions that were
introduced during the long release cycle of 2.0.11. We strongly
recommend upgrading to this version instead of using 2.0.11.

Notable changes include restoring the type and target on unresolved
footnotes (so they display properly), fixing a crash when an extension
set a numeric width or height on an SVG image when using opts=inline,
resetting the word wrap behavior to normal on tables to avoid aggressive
word breaks, reverting the use of a Ruby 2.3 construct that prevented
Asciidoctor from working on CentOS, including the role on an inline
image in the DocBook output, and passing through the explicit width and
height values on an SVG image as is when using opts=inline.

Please note that it will not be possible to use Asciidoctor on Ruby <
2.3 after the 2.0.x release line. Those versions are already
unsupported. However, since we had not yet made a change that prevented
them from working, we'll wait until
2.1.x to do so.

v2.0.11 @mojavelinux released this on Nov 3 2020

This is a big patch release that fixes a slew of issues that have been
discovered from heavy usage. It's a larger release than we had planned
because we're in the process of rewriting the docs and have been holding
minor releases until the docs were merged. Unfortunately, that has taken
longer than expected. But we're nearly there. In the meantime, we're
putting out this patch release to publish the fixes that have
accumulated in the meantime.

Notable changes include fixing two infinite loop scenarios, applying
text formatting to cells in the implicit table head row when the column
has the "a" or "l" style, restoring compatibility with Rouge >= 3.4 (by
defining entry point API methods as class methods instead of module
functions), fix bottom margin collapsing on an AsciiDoc table cell,
removing excess hard line breaks in multi-line AsciiMath blocks,
preserving repeating blackslashes in manpage output, move style tags for
syntax highlighters into head of HTML output, for wrapping to avoid
overflow in HTML output, fix cell borders for colspan or rowspan cells
when frame and grid attribute values are congruent, and allow the
encoding of the include file to be specified using the encoding
attribute.