Add git-multimail version 1.5.0
git-multimail is a tool for sending notification emails on pushes
to a Git repository. It includes a Python module called git_multimail.py,
which can either be used as a hook script directly or can be imported
as a Python module into another script.

git-multimail is derived from the Git project's old
contrib/hooks/post-receive-email, and is mostly compatible with
that script.