aldumb: import aldumb-2.0.3 as wip/aldumb
wip/aldumb provides the allegro libraries for dumb,
i.e. it builds parts of audio/dumb which are normally
excluded due to the allegro dependency.

Description:

DUMB is a module audio renderer library. It reads module files and outputs
audio that can be dumped to the actual audio playback library.

This is a fork of the original dumb by Ben Davis, and provides the Allegro
libraries.