Drop platform specific binaries.
As far as we know, they are only for Cloud access to the controller.

These binaries are built using libgcc_s.1 and pkgsrc will complain
that this library is missing on the install target if wasn't built
using gcc *even* if these libraries are not native.
This is likely a pkgsrc bug, but we really shouldn't be installing
these anyway.