trader: update to 7.15.
ChangeLog:

Version 7.15 (12th January, 2021)
---------------------------------

Changed the default location where game files are saved from `~/.trader`
to `~/.local/share/trader`, so as to follow the `XDG Base Directory
Specification`__.  If set to an absolute path, the ``XDG_DATA_HOME``
environment variable will override this location.  If the `~/.trader`
directory is already present, it will continue to be used.  Updated the
manual page to suit.

__ https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

Updated the Russian translation to use the uppercase polite singular
forms, as is customary in written documents.

Updated to GNU Gettext 0.20 or later.  Updated to the latest snapshot of
the Gnulib GNU Portability Library.  Also updated the `INSTALL` file to
list the latest tested versions of operating systems, including a new
listing for Haiku R1/beta2.