tex-stix2-otf{,-doc}: import tex-stix2-otf{,-doc}-2.11
The Scientific and Technical Information eXchange (STIX) fonts
are intended to satisfy the demanding needs of authors,
publishers, printers, and others working in the scientific,
medical, and technical fields. They combine a comprehensive
Unicode-based collection of mathematical symbols and alphabets
with a set of text faces suitable for professional publishing.
The fonts are available royalty-free under the SIL Open Font
License.