exabgp4: Initial import
An upgrade to the existing net/exabgp port.  However, since the
configuration format has been changed between ExaBGP 3 and 4, this is a
new port, as not to break running configurations.

This is my first port to pkgsrc, I went to some lengths to fix pkglint
issues, but I was unable to find a replacement for "${CP} should not be
used to install files", so that is still there.

The latest release contains a bug that prevents the code from running on
Python 3.8, a patch is included.  The bug is already fixed upstream.

Additionally, the manpage is patched to account for the prefix, and the
installation script is patched in order to not install things in /etc.