terraform014: Move terraform to terraform014
In order to upgrade to major Terraform version (e.g. from 0.12 to 0.14)
it is needed to upgrade through the latest minor releases of all of the
intermediate versions first (e.g. first to 0.13.x and then 0.14.x).

To be able to do that we will probably need to have a
net/terraform metapackage (or build glue) and then net/terraform012,
net/terraform013 and net/terraform014.