fs2open: update to fs2open-21.0.0
(The major version number denotes the year of release, fs2open-20.0.0 was not released.)

Announcement:

Release: 21.0.0

on: January 27, 2021, 11:33:43 am
Another year, another release!

Change log: (chronologically ordered)
  * More OpenGL optimizations
  * Full type information output for Lua documentation. Very useful for
    writing Lua scripts.
  * A new -weaponspew command-line option for printing MediaVP-style
    statistics, plus a weapon comparison spreadsheet, to the debug log.
  * Improved ship lookup behavior in SEXPs for improved performance.
  * BPTC/BC7 texture compression support
  * Refactoring and new features for the ship lab.
  * Missile multi-lock. It's finally in an official build!
  * Enhancements to the model code in preparation for cool features in
    21.2.
  * IPv6 support for multiplayer
  * A lot of general fixes for multiplayer.

For a summary of changes from a modders point of view, you can check
out EatThePaths summary here: https://www.hard-light.net/forums/
index.php?topic=97094.0

Deprecations:
  * Dropped support for Windows XP
  * Scripting:
      - ba.getFrametime(): The parameter value was used incorrectly
        inside the implementation. To avoid breaking existing scripts
        and to improve readability this has been split into
        ba.getMissionFrametime() and ba.getRealFrametime().
      - gr.drawMonochromeImage(): gr.drawImage() got a new parameter
        for drawing monochrome images and is more flexible in general
        so the monochrome variant is no longer needed.

Deprecations are a mechanism in FSO where a certain feature or aspect
of the engine has changed or is no longer supported. Since this would
normally break existing mods we have the mod table feature "$Target
Version:" with which a mod can specify what version of FSO it was
developed with. The features listed above will be removed or changed
when the target version of a mod is at least the version released in
this post.