coturn: Partial update to 4.5.2
The current version was marked as having vulnerabilities, and since
this is wip, commit partial work on an update to 4.5.2 which might fix
them.

Don't patch away upstream's patch-install.txt, because this is fragile
and if we don't want to install it we can just not install it.

Set LICENSE; this is per the github page just modified-bsd.

This does not build because libevent2 is not found.

Enhance TODO:
  Note that COMMIT_MSG is needed.
  Note that doc install path needs updating
  Note that MESSAGE should be removed (and any doc needed installed instead)