edbrowse: Update to 3.7.5.1
pkgsrc changes:
 - Remove patches/patch-src_http.c, now present in the release.

Changes:
========
3.7.5.1
-------
Search function on gopher.

Use threads to download files in the background and in parallel,
instead of processes.

Get the windows port working again; it hasn't worked for a while.
It didn't work because of fork; now that we are threaded it can comne back to life.
That by itself seems worth releasing a mini-version.

3.7.5
-----
ib command to allocate input buffers for textareas.

fll command to set the format line length. fll 80+ to allow a little overflow past 80 columns.

Convert dos \r\n newlines to simple \n.
This always happened in the dos world, now it happens in unix too.
Writing the same file back out converts newlines back.
It can be controlled by the iu command.
This follows the model of iso8859-1 to and from utf8.

Download javascript files in parallel and in background, experimental. Toggle with jsbg command. Works with openssl but not gnutls.

trace and breakpoints in the deminimized local javascript files.

snapshot() function to build a local copy of a website for debugging.

Don't include the (substantial) deminimization code unless EBDEMIN is set.