ats2: Update wip/ats2 to 0.4.2
* 0.4.2

This is the 53rd release of ATS2, the successor of the
ATS programming language. The compiler for ATS2 is given
the name ATS/Positats, ATS2/Postiats or simply Postiats.

The official website for ATS is:

http://www.ats-lang.org

ATS-Postiats is hosted at github:

https://github.com/githwxi/ATS-Postiats

Major releases of ATS2 are available at:

https://sourceforge.net/projects/ats2-lang/

Major releases of external packages for ATS2 are available at:

https://sourceforge.net/projects/ats2-lang-contrib/

Here is a list of major additions and changes since the last release:

There are only a few minor modifications, which are primarily
needed for compiling the current (2020-11-01) implementation of ATS.

1. Fixing the handling of an issue causing the compiler crash:
   the proof of the at-view of a local variable is not restored
   at the end of its scope.
   Please visit the following link for more details:
   https://groups.google.com/g/ats-lang-users/c/f9QVilhHMEM
2. Fixing bug-2020-10-14
   (for the omission of handling records when the exhaustiveness of
    pattern matching is tested)
   Kudos to Alex Dambaev for reporting it!

Please note that in this release, there are two variants of
the package released: The "gmp" variant depends on the GMP library and
the "int" variant does not depend on the GMP library. In this release
ATS2-Postiats-0.4.2.tgz is just the same as ATS2-Postiats-int-0.4.2.tgz.

----

* 0.4.1

This is the 52nd release of ATS2, the successor of the
ATS programming language. The compiler for ATS2 is given
the name ATS/Positats, ATS2/Postiats or simply Postiats.

The official website for ATS is:

http://www.ats-lang.org

ATS-Postiats is hosted at github:

https://github.com/githwxi/ATS-Postiats

Major releases of ATS2 are available at:

https://sourceforge.net/projects/ats2-lang/

Major releases of external packages for ATS2 are available at:

https://sourceforge.net/projects/ats2-lang-contrib/

Here is a list of major additions and changes since the last release:

There are only a few minor modifications, which may be needed for
compiling the implementation of ATS3 in the future.

Please note that starting in this release, there are two variants of
the package released: The "gmp" variant depends on the GMP library and
the "int" variant does not depend on the GMP library. In this release
ATS2-Postiats-0.4.1.tgz is just the same as
ATS2-Postiats-int-0.4.1.tgz.