qt6-qtbase: proper -fp-model when gcc fix
add patch that fixes the problem that the -fp-model switch gets added
when gcc is used. The line that sets ICC checks for icl or icc in
CMAKE_C_COMPILER and it can include the whole path to the compiler.
this causes ICC to be set if the pkgsrc directory is in a path that
has a directory with a name that includes icc or icl.