openblas: flags from environment again to un-break PICy build
The bug that prompted us to force the compiler flags in the
make arguments is fixed. Now we got a different one: The logic
that decides to add -fPIC where needed is overridden when doing
this, resulting in relocation errors (strangely, not with every
toolchain).

So let's remove that again and take FFLAGS and friends from the
environment again.