qt6-qtbase: add fixes for FreeBSD
Add fixes and PLIST for FreeBSD.
Put gssapi in findgssapi first since FreeBSD has both gssapi and gsssapi_krb5
but the functions are in gssapi.
Add --warn-unresolved-symbols to ldflags since environ is not a waek symbol in
libc and that breaks the build.
The patch for configure got updated automaticly as well.