qt5-qtwebengine: version 5.15.2
based on the OpenBSD qt5-qtwebengine port and the pkgsrc-wip chromium.
Builds for me on NetBSD 9.1_STABLE/amd with gcc if you build libxml2
with the icu option and you use libjpeg-turbo as your jpeg package.

Some of the NetBSD specific stuff is a quick hack to get the overall
package building and needs to be revisited.