terraform: Add a terraform meta package
Mostly inspired and several parts completely (and shamelessy!)
copypasted from lang/go.

This is intended to be reused by other wip/terraform{013,014} (and
hopefully a wip/terraform012 soon too!) packages and to store possible
glue for providers and probably to make several version of terraform
cohexist so they can be gradually migrated/updated.