terraform: Add initial PoC provider.mk
Intended to be used by all Terraform providers (still WIP-ish, needs to be
properly documented and then discussed and a symbolic link to bin/ should be
probably installed as well because Terraform 0.12 needs it (and Terraform 0.13
and 0.14 needs to be patched in order to honor the directory in pkgsrc ${PREFIX}
too!)).

XXX: The note about `do-install:' is probably a bit error-prone... is there any
XXX: better way to always overwrite the go-{module,package}.mk one when
XXX: terraform/provider.mk is included?